Rename test-only convertor file
